+++ openghg_inversions/get_data.py
@@ -41,7 +41,7 @@ def data_processing_surface_notracer(
     inlet=None,
     instrument=None,
     calibration_scale=None,
-    met_model=None,
+    met_model: Optional[list] = None,
     fp_model=None,
     fp_height=None,
     fp_species=None,
@@ -96,8 +96,8 @@ def data_processing_surface_notracer(
             (length must match number of sites)
         calibration_scale (str):
             Convert measurements to defined calibration scale
-        met_model (str/opt):
-            Meteorological model used in the LPDM.
+        met_model (list/opt):
+            Meteorological model used in the LPDM. List must be same length as number of sites.
         fp_model (str):
             LPDM used for generating footprints.
         fp_height (list/str):
@@ -143,6 +143,8 @@ def data_processing_surface_notracer(
         fp_height = [None] * nsites
     if obs_data_level is None:
         obs_data_level = [None] * nsites
+    if met_model is None:
+        met_model = [None] * nsites
 
     fp_all = {}
     fp_all[".species"] = species.upper()
@@ -207,6 +209,7 @@ def data_processing_surface_notracer(
                 height=fp_height[i],
                 domain=domain,
                 model=fp_model,
+                met_model=met_model[i],
                 start_date=start_date,
                 end_date=end_date,
                 store=footprint_store,